+++ engine/include/fury/fury.hpp
@@ -41,7 +41,7 @@ namespace fury {
 
 inline constexpr const char* engine_name() { return "Fury"; }
 inline constexpr int version_major() { return 2; }
-inline constexpr int version_minor() { return 7; }
+inline constexpr int version_minor() { return 8; }
 inline constexpr int version_patch() { return 0; }
 
 }  // namespace fury